requirements.txt
# Root development environment (training + local tests)
# Separate deployment-specific requirement files already exist under
# azure_recommender/functions_reco/  and  azure_recommender/frontend_web/

numpy>=1.23,<1.26
pandas>=2.0,<2.2
lightgbm==3.3.5  # training & inference
scikit-learn==1.5.0   # optional: scaling, metrics
requests>=2.31
streamlit==1.35.0
azure-functions==1.13.*  # needed only when running the Function locally
azure-identity>=1.14  # optional: Azure auth helpers
tqdm>=4.66
